third_party/WebKit/LayoutTests/fast/css/parsing-webkit-font-smoothing-expected.txt

This tests checks that all of the input values for -webkit-font-smoothing parse correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test("-webkit-font-smoothing: auto;") is "auto"
PASS test("-webkit-font-smoothing: none;") is "none"
PASS test("-webkit-font-smoothing: antialiased;") is "antialiased"
PASS test("-webkit-font-smoothing: subpixel-antialiased;") is "subpixel-antialiased"
PASS test("-webkit-font-smoothing: apple;") is ""
PASS test("-webkit-font-smoothing: 15;") is ""
PASS test("-webkit-font-smoothing: auto auto;") is ""
PASS successfullyParsed is true

TEST COMPLETE
